HACK: arm: tegra: panel: fix Jetson fb0 setup by choosing dummy LCD config
To use HDMI as the secondary display and allow for an LCD to be attached to
Jetson's expansion PINS, we need to specify an LCD during panel init.  This
fixes fb0 setup and stops Nvidia's hwcomposer from exploding.

Randomly chose a 1920x1200 display: p,wuxga-10-1 (also included in dts)

Later, this can be adjusted for whatever LCD is to be tested.

Change-Id: Ibebdaa2c82855c8f21e88a3a80e823358bfca9c8
Signed-off-by: Michael Scott <michael.scott@linaro.org>